fix: 修复 assembleRelease 下两处编译错误（postDelayed 延迟参数类型 + maxFramerate 类型）+ 记录 so 页对齐安装排查经验 Co-authored-by: monkeycode-ai <monkeycode-ai@chaitin.com>

+++ app/src/main/java/com/screenshare/MainActivity.kt
@@ -1142,7 +1142,7 @@ class MainActivity : AppCompatActivity(), WebRTCPeer.Listener {
                 (android.media.ToneGenerator.MAX_VOLUME * 0.5).toInt()
             )
             tg.startTone(toneType, durationMs)
-            binding.root.postDelayed({ try { tg.release() } catch (_: Throwable) {} }, durationMs + 200)
+            binding.root.postDelayed({ try { tg.release() } catch (_: Throwable) {} }, (durationMs + 200).toLong())
         } catch (t: Throwable) {
             Log.w(TAG, "提示音播放失败: ${t.message}")
         }
